requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.7
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-tempus-dominus==5.1.2.17
django-timezone-field==5.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
